combustion, config under rails 5.0 to avoid deprecation notice
choose the older behavior.

DEPRECATION WARNING: Time columns will become time zone aware in Rails 5.1. This
still causes s to be parsed as if they were in ,
and s to be converted to .
To keep the old behavior, you must add the following to your initializer:
    config.active_record.time_zone_aware_types = [:datetime]
To use the new behavior, add the following:
    config.active_record.time_zone_aware_types = [:datetime, :time]