raise on bad model cast
We had been trying to return nil, since this seemed consistent with existing ActiveRecord primitive values, which kind of go to some base value on a bad cast.

But this ended up making errors in our real client code confusing and hard to debug. Let's try raising early.

Note this means we'll raise on trying to fetch something from the DB with data that can't be cast too, we may have to revisit that.