move query_method implementation to module method
So as not to pollute the class that is including AttrJson::Record with an instance method it doesn't need.

Sadly we could not re-use Rails code here, becuase the built-in method assumes attribute
can be obtained with `self[attr_name]`, which you can not with attr_json (is that bad?), as
well as `self.class.columns_hash[attr_name]` which you definitely can not (which is probably not bad),
and has no way to use the value-translation semantics independently of that. May be a problem if
ActiveRecord changes it's query method semantics in the future, will have to be sync'd here.