test under MRI 2.6.5 too, for Rails 5.2+
We don't want to REMOVE any old ruby versions without a major version release, cause we want to promise it still works under those. So we have a lot of versions tested under now. But we'll only add the new one for Rails 5.2+, since earlier versions are out of support for Rails team anyway.

No need to test github edge Rails on anything but most recent MRI -- we're making no promises about what versions of MRI we support on that one!