AttrJson::Model#== requires exact same class, not sub-class
Better semantics. Avoids potential for bugs like the one @caiofilipemr  kindly reported in #84. Where certain other classes compared for equality would raise.

Also now with tests of the == method.

While if we're being honest this is technically a backwards compat, we're going to call it a bug fix, that didn't make any sense, and I doubt anyone was relying on it.