polymorphic single type can be set to nil
It behaved inconsistently before, where it could have a nil original value, but once you set it to something else, you couldn't set it back without an unexpected raise.

This seems most consistent/backwards-compat, to let you manually set it to nil.

This ends up meaning that a polymorphic array type lets you do , where before it would raise unexpectedly. Not sure what we actually want it to do there, but this is most like a backwards-compat no-change at present, fixing the bug with an unexpected raise. See also #114 questioning in general what is expected behavior with array types and nil