Handle coffeescript dependency weirdness
Even though we don't use coffee-script, when running specs, some part of rails
or other part of our stack is still insisting on requiring it, for reasons we
don't understand, so we need to depend on it.

We want ./bin/rspec to work without involving appraisal, the appraisal setup should be optional. So we move this oddly necessary coffeescript dependency to Gemfile instead of just Appraisal.