explicitly include sprockets to get Rails-edge test to pass
We don't actually use sprockets at all, but the Combustion gem we use to help with testing tries to init it for Rails. To get our tests to pass, we just make Combustion happy in the rails 7-edge environment by adding sprockets-rails as a dependency. https://github.com/pat/combustion/issues/119