include sprockets-rails in dependencies for rails 7.0.0 tests
Not sure if this is really the best way to test, but it gets tests to pass for now. We're trying to test cocoon works with our attributes, and I guess cocoon is currently only doing sprockets though... so I guess it is best.

Maybe we should include our own cocoon-like JS via some other method in the future... https://github.com/nathanvda/cocoon/issues/555