Fix bug involving conflict between AttrJson::Model NestedAttributes and defaults
Ran into this bug in a real world case involving setting array default to empty array. It was very confusing and complicated, I found it too hard to reproduce the actual case in a spec, this simple spec was the best I could do for failing before the change passing after.

Only apply defaults after initialiation, with args passed in initialization, only if defaults are still needed. Previously, defaults were sometimes accidentally over-writing values set with `*_attributes` keys, in real world use cases.

This is a better way of doing defaults anyway... not sure what the comment saying that method couldn't be private was about, or why it was on class instead of instance. I think those were leftover from previous implementations, currently this is simpler anyhow.