ensure serializes to UTC, with tests
We didn't actually have any tests previously that ensured we were serializing to UTC?? Or not enough. We add a bunch of test coverage, and then feel comfortable removing these few custom lines that forced UTC on #serializable_hash -- they weren't necessary, we're getting UTC anyway.

I think because Time#iso8601 method Rails is using to serialize our times to json insists on serialiing in UTC? But now we have tests to ensure outcome is as we want.