Ensure time/datetime are by default cast to ActiveSupport::JSON::Encoding.time_precision
We're using ActiveSupport json serialization, and this is what it will do anyway -- let's match. We were formerly really confused aboout what Rails was doing, but I think Rails got a little bit more predictable since 4.2 when we first wrote this; and I just spent the time to get to the bottom of it and make sure I understood it. So this can now all be quite a bit more clear and predictable.