Fix build for post-Rails-7.1 main branch changes to dirty tracking
We set a default of empty hash on our container attribute.  Up to Rails 7.1, Rails ignores the default in dirty tracking, and consideers the initial value nil -- but post Rails 7.1, Rails dirty tracking _starts_ at `{}`. We don't consider it a bug, the new behavior is probably better, we just go with it in tests either way.