Keep nested model validation working in Rails post 7.1
Without using ActiveRecord::Validations::AssociatedValidator -- we now supply our own custom validation code, based on the same basic thing it was doing.

We had been using this ActiveRecord class in an analogous context for our AttrJson-specific nested model validations. And it used to work, because the code was simple enough. But as a result of https://github.com/jrochkind/attr_json/pull/220 , it started to contain some ActiveRecord-specific logic that broke us in our analagous use case.

No problem, we can just supply validation code ourselves, forked from the last working copy. It's very simple logic.

We even had a note to ourselves in a comment that we might have to do this at some point!