Allow double parens in SQL spec
On Rails post-7.1 edge @ c402ec78724a, SQL generated for certain 'NOT' queries has a double set of parens nested.

Our spec testing for SQL then fails.  We modify it to accept double set of parens.  This makes no semantic difference, our code was working on this version of Rails regardless; the spec just has to accep the variant SQL that means the same thing.

(No idea why Rails changed to generate a double set of parens here, or if it's a bug somewhere; but the SQL still works fine, an unnecessary double set of parens doesn't hurt anything I'm aware of, so I'm not going to worry about it).