Workaround bug where rails does not 'require' logger
Logger got moved to default gems a while ago in ruby versions, and rails should have been 'require'ing it but wasn't. This was hidden by the fact that concurrent-ruby used to 'require' it, but no longer does. With recent versions of concurrent-ruby, somebody needs to 'require' logger -- Rails does not plan to release a fix for pre-7.1 rails.

So we can simply require it in our spec_helper for CI no big deal.